@@ -1,6 +1,6 @@
 module github.com/snicol/rf
 
-go 1.14
+go 1.18
 
 require (
 	github.com/go-chi/chi v4.1.1+incompatible
@@ -8,5 +8,12 @@ require (
 	github.com/sirupsen/logrus v1.6.0
 	github.com/snicol/yael v0.0.0-20200504191458-dca710d34709
 	github.com/xeipuuv/gojsonschema v1.2.0
-	golang.org/x/net v0.0.0-20200505041828-1ed23360d12c // indirect
+)
+
+require (
+	github.com/konsorten/go-windows-terminal-sequences v1.0.3 // indirect
+	github.com/xeipuuv/gojsonpointer v0.0.0-20180127040702-4e3ac2762d5f // indirect
+	github.com/xeipuuv/gojsonreference v0.0.0-20180127040603-bd5ef7bd5415 // indirect
+	golang.org/x/net v0.0.0-20220225172249-27dd8689420f // indirect
+	golang.org/x/sys v0.0.0-20211216021012-1d35b9e2eb4e // indirect
 )
